chore: rename `record::str` to `record::test`
